+++ arch/x86/adler32_avx512_vnni.c
@@ -18,20 +18,20 @@
 #include "adler32_avx512_p.h"
 #include "adler32_avx2_p.h"
 
-Z_INTERNAL uint32_t adler32_avx512_vnni(uint32_t adler, const uint8_t *src, size_t len) {
+Z_INTERNAL uint32_t adler32_avx512_vnni(uint32_t adler, const uint8_t *src, uint64_t len) {
     if (src == NULL) return 1L;
     if (len == 0) return adler;
 
     uint32_t adler0, adler1;
     adler1 = (adler >> 16) & 0xffff;
-    adler0 = adler & 0xffff; 
+    adler0 = adler & 0xffff;
 
 rem_peel:
     if (len < 32)
 #if defined(X86_SSSE3_ADLER32)
         return adler32_ssse3(adler, src, len);
 #else
-        return adler32_len_16(adler0, src, len, adler1); 
+        return adler32_len_16(adler0, src, len, adler1);
 #endif
 
     if (len < 64)
@@ -40,7 +40,7 @@ rem_peel:
 #elif defined(X86_SSE3_ADLER32)
         return adler32_ssse3(adler, src, len);
 #else
-        return adler32_len_16(adler0, src, len, adler1); 
+        return adler32_len_16(adler0, src, len, adler1);
 #endif
 
     const __m512i dot2v = _mm512_set_epi8(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
@@ -54,7 +54,7 @@ rem_peel:
     while (len >= 64) {
         vs1 = _mm512_zextsi128_si512(_mm_cvtsi32_si128(adler0));
         vs2 = _mm512_zextsi128_si512(_mm_cvtsi32_si128(adler1));
-        size_t k = MIN(len, NMAX);
+        uint64_t k = MIN(len, NMAX);
         k -= k % 64;
         len -= k;
         __m512i vs1_0 = vs1;
@@ -117,16 +117,16 @@ rem_peel:
         goto rem_peel;
     }
 
-    return adler; 
+    return adler;
 }
 
-Z_INTERNAL uint32_t adler32_fold_copy_avx512_vnni(uint32_t adler, uint8_t *dst, const uint8_t *src, size_t len) {
+Z_INTERNAL uint32_t adler32_fold_copy_avx512_vnni(uint32_t adler, uint8_t *dst, const uint8_t *src, uint64_t len) {
     if (src == NULL) return 1L;
     if (len == 0) return adler;
 
     uint32_t adler0, adler1;
     adler1 = (adler >> 16) & 0xffff;
-    adler0 = adler & 0xffff; 
+    adler0 = adler & 0xffff;
 
 rem_peel_copy:
     if (len < 32) {
@@ -138,7 +138,7 @@ rem_peel_copy:
 #if defined(X86_SSSE3_ADLER32)
         return adler32_ssse3(adler, src, len);
 #else
-        return adler32_len_16(adler0, src, len, adler1); 
+        return adler32_len_16(adler0, src, len, adler1);
 #endif
     }
 
@@ -151,7 +151,7 @@ rem_peel_copy:
     while (len >= 32) {
         vs1 = _mm256_zextsi128_si256(_mm_cvtsi32_si128(adler0));
         vs2 = _mm256_zextsi128_si256(_mm_cvtsi32_si128(adler1));
-        size_t k = MIN(len, NMAX);
+        uint64_t k = MIN(len, NMAX);
         k -= k % 32;
         len -= k;
         __m256i vs1_0 = vs1;
@@ -219,7 +219,7 @@ rem_peel_copy:
         goto rem_peel_copy;
     }
 
-    return adler; 
+    return adler;
 }
 
 #endif
